requirements.txt
pyTelegramBotAPI == 4.12.0
python-dotenv == 1.0.0
requests == 2.31.0
peewee == 3.16.2
pydantic == 1.10.9
geopy == 2.3.0
loguru == 0.7.0
